[CHORE] Cleanup and code format
