+++ Example/iOS/Views/LinkView.swift
@@ -50,11 +50,11 @@ final class LinkView: UIView, EpoxyableView {
             return animationView
           },
           setContent: { context, _ in
-              if let animation = LottieAnimation.named(animationName) {
-                  context.constrainable.animation = animation
-              } else if let lottie = DotLottie.named(animationName) {
-                  context.constrainable.dotLottie = lottie
-              }
+            if let animation = LottieAnimation.named(animationName) {
+              context.constrainable.animation = animation
+            } else if let lottie = DotLottie.named(animationName) {
+              context.constrainable.dotLottie = lottie
+            }
             context.constrainable.contentMode = .scaleAspectFit
             context.constrainable.currentProgress = 0.5
           })
