+++ Sources/Private/Model/Extensions/Bundle.swift
@@ -19,22 +19,22 @@ extension Bundle {
     return nil
     #endif
   }
-    
-    func getDotLottieData(_ name: String, subdirectory: String? = nil) throws -> Data? {
-      // Check for files in the bundle at the given path
-      let name = name.removingDotLottieSuffix()
-      if let url = url(forResource: name, withExtension: "lottie", subdirectory: subdirectory) {
-        return try Data(contentsOf: url)
-      }
 
-      // Check for data assets (not available on macOS)
-      #if os(iOS) || os(tvOS) || os(watchOS) || targetEnvironment(macCatalyst)
-      let assetKey = subdirectory != nil ? "\(subdirectory ?? "")/\(name)" : name
-      return NSDataAsset(name: assetKey, bundle: self)?.data
-      #else
-      return nil
-      #endif
+  func getDotLottieData(_ name: String, subdirectory: String? = nil) throws -> Data? {
+    // Check for files in the bundle at the given path
+    let name = name.removingDotLottieSuffix()
+    if let url = url(forResource: name, withExtension: "lottie", subdirectory: subdirectory) {
+      return try Data(contentsOf: url)
     }
+
+    // Check for data assets (not available on macOS)
+    #if os(iOS) || os(tvOS) || os(watchOS) || targetEnvironment(macCatalyst)
+    let assetKey = subdirectory != nil ? "\(subdirectory ?? "")/\(name)" : name
+    return NSDataAsset(name: assetKey, bundle: self)?.data
+    #else
+    return nil
+    #endif
+  }
 }
 
 extension String {
@@ -47,14 +47,14 @@ extension String {
 
     return (self as NSString).deletingPathExtension
   }
-    
-    fileprivate func removingDotLottieSuffix() -> String {
-      // Allow filenames to be passed with a ".json" extension (but not other extensions)
-      // to keep the behavior from Lottie 2.x - instead of failing to load the animation
-      guard hasSuffix(".lottie") else {
-        return self
-      }
 
-      return (self as NSString).deletingPathExtension
+  fileprivate func removingDotLottieSuffix() -> String {
+    // Allow filenames to be passed with a ".json" extension (but not other extensions)
+    // to keep the behavior from Lottie 2.x - instead of failing to load the animation
+    guard hasSuffix(".lottie") else {
+      return self
     }
+
+    return (self as NSString).deletingPathExtension
+  }
 }
