+++ Sources/Public/Animation/LottieAnimation.swift
@@ -157,7 +157,7 @@ public final class LottieAnimation: Codable, DictionaryInitializable {
   /// Markers
   let markers: [Marker]?
   let markerMap: [String: Marker]?
-    
+
   /// DotLottie configuration to setup the player
   var dotLottieConfiguration: DotLottieConfiguration?
 }
